+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#30](https://github.com/LaxarJS/ax-developer-tools-widget/issues/30): added prototype page inspector
+
 
 ## v1.6.0